requirements.txt
torch==2.1.0
torchvision
opencv-python
pycocotools
matplotlib
Pillow
numpy
einops
timm
albumentations
termcolor
tqdm
pandas
typed-argument-parser
pytorch-lightning
streamlit
black
huggingface-hub
